requirements.txt
# FastAPI and Web Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1

# Video Downloading
yt-dlp==2023.11.16

# Image Processing
Pillow==10.1.0

# Document Processing
python-docx==1.1.0
pypdf==3.17.1
openpyxl==3.1.2
pandas==2.1.3
xlrd==2.0.1

# Utilities
python-magic-bin==0.4.14; sys_platform == 'win32'
python-magic==0.4.27; sys_platform != 'win32'

# CORS
starlette==0.27.0
